Analysis with and without surprisal

+++ src/imdb.py
@@ -129,7 +129,7 @@ class SkipRNN():
         # print(f"Vector for unknonw words is {embeddings_index.get('unk')}")
         embedding_matrix = np.zeros((tot_len, self.BATCH_SIZE, self.SEQUENCE_LENGTH, self.EMBEDDING_LENGTH), dtype=np.float32)
         probs_matrix = np.ones((tot_len, self.BATCH_SIZE, self.SEQUENCE_LENGTH, 1), dtype=np.float32)
-        mask = np.zeros((tot_len, self.BATCH_SIZE, self.SEQUENCE_LENGTH, 1))
+        mask = np.zeros((tot_len, self.BATCH_SIZE, self.SEQUENCE_LENGTH, 1), dtype=np.float32)
         labels = np.zeros((tot_len, self.BATCH_SIZE), dtype=np.int64)
         line_index = 0
         batch_index = 0
@@ -176,7 +176,7 @@ class SkipRNN():
                                  name='Samples')  # (batch, time, in)
         ground_truth = tf.placeholder(tf.int64, shape=[self.BATCH_SIZE], name='GroundTruth')
         probs = tf.placeholder(tf.float32, shape=[self.BATCH_SIZE, self.SEQUENCE_LENGTH, 1], name='Probs')
-        mask = tf.placeholder(tf.int32, shape=[self.BATCH_SIZE, self.SEQUENCE_LENGTH, 1], name='padding_mask')
+        mask = tf.placeholder(tf.float32, shape=[self.BATCH_SIZE, self.SEQUENCE_LENGTH, 1], name='padding_mask')
 
         cell, initial_state = create_model(model='skip_lstm',
                                            num_cells=[self.HIDDEN_UNITS],
